Uploaded exact copy of my files from the laptop.
What you are looking at here is the website chatbot tester. There are no libraries needed for this, it should be a "plug and play" bit of code. When running, you should be taken to opening up a web tab, clicking on the bottom right, a small animation so the chatbot pops up, and then can interact. Buttons are pretty useless and gimmicky, but the chatbot should tell you what you can and can't do :)